feat: :sparkles: rename file and add fetch section hook
